requirements.txt
# Core deep learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer vision and video processing
opencv-python>=4.6.0
Pillow>=9.0.0
albumentations>=1.3.0

# Data manipulation
numpy<2.0.0
pandas>=1.4.0
scikit-learn>=1.1.0

# Jupyter and visualization
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment tracking
wandb>=0.13.0

# Utilities
tqdm>=4.64.0
PyYAML>=6.0
requests>=2.28.0

# Optional: GPU support (uncomment if you have CUDA)
# torch>=1.12.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.13.0+cu113 -f https://download.pytorch.org/whl/torch_stable.html 